waldo-twitter-bot/requirements.txt
requests
python-dotenv
Flask
Flask-Cors
Flask-Limiter
xrpl-py
redis>=4.0

